requirements.txt
flask
gunicorn
psycopg2-binary
webauthn
python-dotenv
SQLAlchemy
alembic
Flask-SQLAlchemy
Flask-Migrate
flask-login